Changing hierarchy + add routes on App component
