+++ src/components/MainPage/MainPage.css
@@ -1,0 +1,5 @@
+header {
+  display: flex;
+  align-items: center;
+  padding-left: 40px;
+}
