+++ src/components/PostPage/PostPage.tsx
@@ -1,0 +1,7 @@
+export const PostPage = () => {
+  return (
+    <>Hello</>
+  );
+}
+
+export default PostPage;
